Replaces AuthLib with Custom library
AuthLib continuously threw schema errors that I could not solve.

This is 100% due to my lack of familiarity with how to interact with the Instagram endpoints.

Created a new custom Instagram API which uses very simple api calls to achieve the current proof of concept I was looking for.

This removed the OAuth token exchange and authorization, because I don't need them right now.

This needs to be used alongside that social_apis repository that I will be committing shortly